+++ test/lsp/analyzer_test.dart
@@ -54,6 +54,7 @@ int one() { return 1; }
     expect(Analyzer.languageOf('file:///a.dart'), 'dart');
     expect(Analyzer.languageOf('file:///a.java'), 'java');
     expect(Analyzer.languageOf('file:///a.go'), 'go');
+    expect(Analyzer.languageOf('file:///a.apollo'), 'apollo');
     expect(Analyzer.languageOf('file:///a.unknown'), isNull);
   });
 }